Use Factory instead of DIC
Using the dependency injection container was an experiment, to see how
the code would look like. In the test environment, it is better mot to
use one, as a factory provides a more clear interface. Having a a shared
object that is passend around is solved by the getSharedObject function.