Fix Validations
After switching to this repo on FundraisingFrontend, some E2E tests of
FF failed, this is now fixing those failing tests and adding new tests
that were not covered before and should not affect the E2E test in
FundraisingFrontend.

Some tests and fixtures from FundraisingFrontend, that were not in
the "contexts" directory from which this repo was created, were moved
into this repo.

Added the "fix-cs" Makefile target.