Modernize Doctrine initialization
Implement more modern Doctrine initialization in
TestSubscriptionContextFactory, removing deprecated and dropped methods

Change interface of SubscriptionContextFactory. This is a BC break, but
the dependents of this bounded context are ready for it, so not breakage
will occur when (auto)updating to a new release.

Ticket: https://phabricator.wikimedia.org/T312080